+++ lib/muffin_man.rb
@@ -19,6 +19,7 @@ require "muffin_man/fba_inventory/v1"
 require "muffin_man/request_helpers"
 require "muffin_man/feeds/v20210630"
 require "muffin_man/notifications/v1"
+require "muffin_man/merchant_fulfillment/v0"
 
 module MuffinMan
   class Error < StandardError; end